+++ api/getlinks.js
@@ -13,15 +13,17 @@ export default async function handler(req, res) {
 
     //Delete the entry if its expired and return error
     if (!session || session.expiresAt < new Date()) {
-        await prisma.session.delete({
-            where: { token }
-        });
+        if (session) {
+            await prisma.session.delete({
+                where: { token }
+            });
+        }
 
         return res.status(401).json();
     }
 
     //return all links to the user when authenticated
-    const urls = await prisma.url.findMany({orderBy: { createdAt: "desc" }});
+    const urls = await prisma.url.findMany({ orderBy: { createdAt: "desc" } });
 
     return res.status(200).json(urls);
 }
