+++ api/verify.js
@@ -13,9 +13,11 @@ export default async function handler(req, res) {
 
     //Delete the entry if its expired and return valid: false to the user
     if (!session || session.expiresAt < new Date()) {
-        await prisma.session.delete({
-            where: { token }
-        });
+        if (session) {
+            await prisma.session.delete({
+                where: { token }
+            });
+        }
 
         return res.status(401).json({ valid: false });
     }
